Refs #6007. Add rmw_guard_condition.c file.
