+++ rmw_microxrcedds_c/src/rmw_guard_condition.c
@@ -1,0 +1,38 @@
+// Copyright 2019 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "utils.h"
+
+#include <rmw/rmw.h>
+#include <rmw/allocators.h>
+
+rmw_guard_condition_t *
+rmw_create_guard_condition(rmw_context_t * context)
+{
+  EPROS_PRINT_TRACE()
+
+  rmw_guard_condition_t * rmw_guard_condition = (rmw_guard_condition_t *)rmw_allocate(
+    sizeof(rmw_guard_condition_t));
+
+  return rmw_guard_condition;
+}
+
+rmw_ret_t rmw_destroy_guard_condition(rmw_guard_condition_t * guard_condition)
+{
+  EPROS_PRINT_TRACE()
+
+  rmw_free(guard_condition);
+
+  return RMW_RET_OK;
+}
